+++ src/lib/engn/entity.rs
@@ -1,14 +1,9 @@
 use {
-    super::{
-        ray::{
-            Ray, InceptedRays,
-        }
-    },
     crate::{
         errs::{
             ReRes,
             ReErr::{self, *},
-            EngnErr::{self, *},
+            GameErr::{self, *},
             GridErr::{self, *},
             MathErr::{self, *},
         },
@@ -29,14 +24,13 @@ use {
             Any,
             TypeId
         },
-        f64::consts::PI,
     },
     uuid::Uuid,
 };
 
 
 /// Matrix of `Uuid` (standard v4) allocated in heap
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug, Default, Clone, PartialEq)]
 pub struct IdPool {
     ids: Vec<Rc<Uuid>>,
 }
@@ -79,6 +73,8 @@ pub enum Prop {
     Pos,
     /// Direction of view, `Matrix`
     Dir,
+    /// Roll of entity, `Matrix`
+    Roll,
     /// Field-of-view, `f64`
     Fov,
     /// Vertical field-of-view, `f64`
@@ -100,11 +96,11 @@ impl Property for Prop {
 #[derive(Debug)]
 pub struct EntityCore {
     /// Cloned `Rc` from actual `Game` instance
-    cs: Rc<CoordSys>,
+    pub(in super) cs: Rc<CoordSys>,
     /// Cloned `Rc` from `IdPool` within actual `Game` instance
-    id: Rc<Uuid>,
+    pub(in super) id: Rc<Uuid>,
     /// Dictionary with `(TypeId, isize)` as key and `Any` trait object wrapped into `Box` as value
-    props: HashMap<(TypeId, isize), Box<dyn Any>>,
+    pub(in super) props: HashMap<(TypeId, isize), Box<dyn Any>>,
 }
 
 impl EntityCore {
@@ -132,7 +128,7 @@ impl EntityCore {
         if let Some(prop) = self.props.get(&key.feed()) {
             Ok(prop)
         } else {
-            Err(EngnErr(NotInitializedProp))
+            Err(GameErr(NotInitializedProp))
         }
     }
 
@@ -168,54 +164,14 @@ impl std::fmt::Debug for dyn Entity {
 }
 
 
-/// List of entities implementing shared interior mutability
-#[derive(Debug)]
-pub struct EntityList {
-    entities: Vec<Rc<RefCell<dyn Entity>>>,
-}
-
-impl EntityList {
-    /// Instantiates empty list
-    pub fn new() -> Self {
-        Self { entities: vec![] }
-    }
-
-    /// Appends new entity that must implement Entity
-    pub fn append(&mut self, entity: impl Entity + 'static) {
-        self.entities.push(Rc::new(RefCell::new(entity)));
-    }
-
-    /// Removes entity from the list with the given `Uuid`
-    pub fn remove(&mut self, id: &Rc<Uuid>) {
-        self.entities.retain(|entity| Rc::ptr_eq(&entity.borrow().core().id, id));
-    }
-
-    /// Returns shared interior mutable ref to entity if exists
-    pub fn get(&self, id: &Rc<Uuid>) -> Option<Rc<RefCell<dyn Entity>>> {
-        if let Some(rc) = self.entities.iter().find(|entity| Rc::ptr_eq(&entity.borrow().core().id, id)) {
-            Some(Rc::clone(rc))
-        } else {
-            None
-        }
-    }
-
-    /// Permorms closure that may be immutable due to interior mutability
-    pub fn exec(&self, f: fn(&RefCell<dyn Entity>)) {
-        for entity in self.entities.iter() {
-            f(entity);
-        }
-    }
-}
-
-
 /// Basic game object
 #[derive(Debug)]
 pub struct GameObject {
-    core: EntityCore,
+    pub(in super) core: EntityCore,
 }
 
 impl GameObject {
-    /// Constructor that takes `EntityCore`, position, direction, and then sets
+    /// Constructor that takes `EntityCore`, position, direction, and then glob
     /// such properties to the given core
     pub(in super) fn new(mut core: EntityCore, pos: Point, dir: Matrix) -> Self {
         core.set_prop(Prop::Pos, Box::new(pos));
@@ -249,87 +205,5 @@ impl Entity for GameObject {
 
     fn core_mut(&mut self) -> &mut EntityCore {
         &mut self.core
-    }
-}
-
-
-/// Game camera that contains `GameObject`
-#[derive(Debug)]
-pub struct Camera {
-    go: GameObject,
-}
-
-impl Camera {
-    /// Constructs new camera from the given game object
-    pub fn new(mut game_object: GameObject) -> Self {
-        Self { go: game_object }
-    }
-
-    /// Setting property `Fov` and validating it
-    pub fn set_fov(mut self, fov: f64) -> ReRes<Self> {
-        if fov < 0.0 || PI <= fov {
-            return Err(EngnErr(InvalidPropF64{ key: Prop::Fov, val: fov }))
-        }
-        self.go.core.set_prop(Prop::Fov, Box::new(fov));
-        Ok(self)
-    }
-
-    /// Getting value of property `Fov` if it set or meaningful error
-    pub fn get_fov(mut self) -> ReRes<f64> {
-        Ok(*self.go.core.get_prop(Prop::Fov)?.downcast_ref::<f64>().unwrap())
-    }
-
-    /// Setting property `VFov` and validating it
-    pub fn set_vfov(mut self, vfov: f64) -> ReRes<Self> {
-        if vfov < 0.0 || PI <= vfov {
-            return Err(EngnErr(InvalidPropF64{ key: Prop::VFov, val: vfov }))
-        }
-        self.go.core.set_prop(Prop::VFov, Box::new(vfov));
-        Ok(self)
-    }
-
-    /// Getting value of property `VFov` if it set or meaningful error
-    pub fn get_vfov(mut self) -> ReRes<f64> {
-        Ok(*self.go.core.get_prop(Prop::VFov)?.downcast_ref::<f64>().unwrap())
-    }
-
-    /// Setting property `VFov` and validating it
-    pub fn set_draw_dist(mut self, draw_dist: f64) -> ReRes<Self> {
-        if draw_dist < 0.0 {
-            return Err(EngnErr(InvalidPropF64{ key: Prop::DrawDist, val: draw_dist }))
-        }
-        self.go.core.set_prop(Prop::DrawDist, Box::new(draw_dist));
-        Ok(self)
-    }
-
-    /// Getting value of property `DrawDist` if it set or meaningful error
-    pub fn get_draw_dist(mut self) -> ReRes<f64> {
-        Ok(*self.go.core.get_prop(Prop::DrawDist)?.downcast_ref::<f64>().unwrap())
-    }
-
-    /// Setting property `LookAt`
-    pub fn set_lookat(mut self, lookat: Point) -> Self {
-        self.go.core.set_prop(Prop::LookAt, Box::new(lookat));
-        self
-    }
-
-    /// Getting value of property `LookAt` if it set or meaningful error
-    pub fn get_lookat(mut self) -> ReRes<f64> {
-        Ok(*self.go.core.get_prop(Prop::LookAt)?.downcast_ref::<f64>().unwrap())
-    }
-
-    /// Constructor for bunch of rays having one inception
-    pub fn incepted_rays(h: i32, w: i32) -> InceptedRays {
-        todo!()
-    }
-}
-
-impl Entity for Camera {
-    fn core(&self) -> &EntityCore {
-        &self.go.core
-    }
-
-    fn core_mut(&mut self) -> &mut EntityCore {
-        &mut self.go.core
     }
 }