+++ src/lib/engn/game.rs
@@ -1,9 +1,9 @@
 use {
-    super::{
-        Ray, IdPool, EntityCore, EntityList, GameObject, Camera,
-    },
-    crate::math::{
-        Point, Matrix, CoordSys
+    super::*,
+    crate::{
+        conf::Conf,
+        math::*,
+        math::matrix::set_biform_vec,
     },
     std::{
         rc::Rc,
@@ -18,19 +18,17 @@ use {
 /// Struct responsible for storing current CoordSys and EntityList and running related scripts
 #[derive(Debug)]
 pub struct Game {
-    cs: Rc<CoordSys>,
-    id_pool: IdPool,
-    entities: EntityList,
+    pub(in super) cs: Rc<CoordSys>,
+    pub(in super) id_pool: IdPool,
+    pub(in super) entities: EntityList,
+    pub(in super) canvas: Canvas,
+    pub(in super) camera: Camera,
 }
 
 impl Game {
     /// Constructor that takes CoordSys
-    pub fn new(cs: CoordSys) -> Self {
-        Self {
-            cs: Rc::new(cs),
-            id_pool: IdPool::new(),
-            entities: EntityList::new(),
-        }
+    pub fn new(conf: Conf) -> Self {
+        Self::from(conf)
     }
 
     pub fn run() {
@@ -55,8 +53,50 @@ impl Game {
         EntityCore::new(&self.cs, &self.id_pool.generate())
     }
 
-    /// `GameObject` in current game, calls `self.entity_core()`
+    /// `GameObject` in current game, uses `self.entity_core()`
     pub fn game_object(&mut self, pos: Point, dir: Matrix) -> GameObject {
         GameObject::new(self.entity_core(), pos, dir)
+    }
+
+    /// `Camera` in current game, uses `self.game_object()`
+    pub fn camera(&self) -> &Camera {
+        &self.camera
+    }
+}
+
+impl Default for Game {
+    fn default() -> Self {
+        let conf = Conf::default();
+        Self::from(conf)
+    }
+}
+
+impl From<Conf> for Game {
+    fn from(conf: Conf) -> Self {
+        set_biform(conf.biform);
+        let cs = Rc::new(
+            CoordSys::new(conf.initpt.clone(),
+                          VectorSpace::new(conf.basis).unwrap())
+                .unwrap());
+        let mut id_pool = IdPool::new();
+        let entities = EntityList::new();
+        let canvas = Canvas::new(
+            GameObject::new(
+                EntityCore::new(&cs, &id_pool.generate()), conf.initpt.clone(), conf.camera_dir.clone()),
+            conf.scr_height, conf.scr_width);
+        let camera = Camera::new(
+            GameObject::new(
+                EntityCore::new(&cs, &id_pool.generate()), conf.initpt, conf.camera_dir))
+            .set_fov(conf.camera_fov).unwrap()
+            .set_draw_dist(conf.draw_dist).unwrap()
+            .set_roll(conf.camera_roll).unwrap();
+
+        Self {
+            cs,
+            id_pool,
+            entities,
+            canvas,
+            camera,
+        }
     }
 }