+++ src/lib/errs/math_errs.rs
@@ -35,4 +35,13 @@ pub enum MathErr {
 
     #[error("trying to create rotation matrix from {0} to {0} axis")]
     RotationInOneAxis(usize),
+
+    #[error("trying to set global DIM variable to 0")]
+    ZeroDimSpace,
+
+    #[error("point and vector has different dimensions: {pt_dim:?} and {vec_dim:?}")]
+    PtVecDimMismatch {
+        pt_dim: usize,
+        vec_dim: usize,
+    },
 }