+++ src/lib/lib.rs
@@ -1,5 +1,6 @@
 #![allow(warnings)]
 
+pub mod conf;
 pub mod engn;
 pub mod errs;
 pub mod grid;